Organize vault tests cases.
