+++ contracts/vault/tests/tests/mod.rs
@@ -1,3 +1,7 @@
 use mars_testing::multitest::helpers;
 
-mod test_managed_vault;
+mod vault_helpers;
+
+mod test_binding;
+mod test_deposit_redeem;
+mod test_instantiate;
