+++ contracts/vault/tests/tests/test_instantiate.rs
@@ -1,0 +1,82 @@
+use cosmwasm_std::{coin, Addr};
+use cw_multi_test::Executor;
+use mars_vault::msg::{InstantiateMsg, VaultInfoResponseExt};
+
+use super::helpers::{mock_managed_vault_contract, AccountToFund, MockEnv};
+use crate::tests::{helpers::deploy_managed_vault, vault_helpers::query_vault_info};
+
+#[test]
+fn instantiate_with_empty_metadata() {
+    let fund_manager = Addr::unchecked("fund-manager");
+    let mut mock = MockEnv::new()
+        .fund_account(AccountToFund {
+            addr: fund_manager.clone(),
+            funds: vec![coin(1_000_000_000, "untrn")],
+        })
+        .build()
+        .unwrap();
+    let credit_manager = mock.rover.clone();
+
+    let managed_vault_addr = deploy_managed_vault(&mut mock.app, &fund_manager, &credit_manager);
+
+    let vault_info_res = query_vault_info(&mock, &managed_vault_addr);
+    assert_eq!(
+        vault_info_res,
+        VaultInfoResponseExt {
+            base_token: "uusdc".to_string(),
+            vault_token: format!("factory/{}/vault", managed_vault_addr),
+            title: None,
+            subtitle: None,
+            description: None,
+            credit_manager: credit_manager.to_string(),
+            vault_account_id: None,
+        }
+    )
+}
+
+#[test]
+fn instantiate_with_metadata() {
+    let fund_manager = Addr::unchecked("fund-manager");
+    let mut mock = MockEnv::new()
+        .fund_account(AccountToFund {
+            addr: fund_manager.clone(),
+            funds: vec![coin(1_000_000_000, "untrn")],
+        })
+        .build()
+        .unwrap();
+    let credit_manager = mock.rover.clone();
+
+    let contract_code_id = mock.app.store_code(mock_managed_vault_contract());
+    let managed_vault_addr = mock
+        .app
+        .instantiate_contract(
+            contract_code_id,
+            fund_manager,
+            &InstantiateMsg {
+                base_token: "uusdc".to_string(),
+                vault_token_subdenom: "fund".to_string(),
+                title: Some("random TITLE".to_string()),
+                subtitle: Some("random subTITLE".to_string()),
+                description: Some("The vault manages others money !!!".to_string()),
+                credit_manager: credit_manager.to_string(),
+            },
+            &[coin(10_000_000, "untrn")], // Token Factory fee for minting new denom. Configured in the Token Factory module in `mars-testing` package.
+            "mock-managed-vault",
+            None,
+        )
+        .unwrap();
+
+    let vault_info_res = query_vault_info(&mock, &managed_vault_addr);
+    assert_eq!(
+        vault_info_res,
+        VaultInfoResponseExt {
+            base_token: "uusdc".to_string(),
+            vault_token: format!("factory/{}/fund", managed_vault_addr),
+            title: Some("random TITLE".to_string()),
+            subtitle: Some("random subTITLE".to_string()),
+            description: Some("The vault manages others money !!!".to_string()),
+            credit_manager: credit_manager.to_string(),
+            vault_account_id: None,
+        }
+    )
+}
